Add services as a subtype of ModuleData
Services are modules that act like a pseudo-user and respond to commands.
They come "out of the box" with some useful functionality and shortcuts,
see service.Service for details.
Services are subclasses of ModuleData (instead of being their own seperate class)
because they need to make use of a number of module-level hooks, and it makes it easier
for the implementor if this is done automatically.

What is not implemented here:
    How to determine if a user is a service admin.
        Hopefully to be done later.
    Anything to do with cross-server management of services.