install: Be more selective about what we wipe when re-installing
We want to completely replace code directories (as opposed to a simple overwrite,
this won't cause problems when a file is no longer present or renamed).
We don't want to replace the config file.
We don't want to delete other stuff in the existing install dir, eg. auction logs.
We don't want to re-install dependencies unless we have to.