Refactor LUSERS command to count every time.
The performance penalty is negligible (on the order of 100μs on my machine),
and it clears up all kinds of desync bugs (such as Issue #28).
Note that I changed the format of the high water marks in the storage file,
so they will effectively get reset when you first start using this version.