Replace a lot of magic numbers with configurable equivalents
I kinda took this upon myself because the amount of magic numbers we
have bothered me. Making these configurable is, in my opinion, a change
that will not only make the codebase cleaner and assure that these
numbers are always the same in all cases (see my fix in TOPICLEN for
an example on this one), but also make txircd a lot more configurable
for our users. One could argue that a lot of these numbers should never
be changed, but I prefer going the InspIRCd route of "configurable but
discouraged" over "not configurable at all".